dba: rng: async; code review

+++ src/main/scala/key.scala
@@ -9,4 +9,4 @@ sealed trait Key
 @N(3) final case class EnKey(@N(1) fid: FdKey, @N(2) id: ElKey) extends Key
 
 @N(4) final case class PathKey(@N(1) dir: FdKey, @N(2) name: ElKey) extends Key
-@N(5) final case class ChunkKey(@N(1) path: PathKey, @N(2) idx: Int) extends Key
+@N(5) final case class ChunkKey(@N(1) path: PathKey, @N(2) idx: Long) extends Key
